refactor(manager-dashboard): create reusable BasicProjectInformation Form component
This is part of a WIP to refactor each project type into a separate 'New Project' subform.
